Enable unused_imports, unused_qualifications and trivial_numeric_casts
The first was allowed at the crate root and in four more files; the other
two are allow-by-default. All three are residue of moving C++ files across
rather than anything the Rust needs: 434 imports nothing used, 411 paths
spelled out where the name was already in scope, and 249 casts from a type
to itself.

Most of it is `cargo fix`. Three classes needed hands:

  * `cargo fix --lib` compiles without cfg(test), so it stripped imports
    that only the in-module unit tests reach. Those move into their
    `mod tests` rather than back to the top of the file, where the
    non-test build would call them unused in turn. Same shape in the
    cfg-gated SIMD sets, which the default aarch64 build never compiles:
    x86_64 and wide are now checked per feature and per target.

  * `trivial_numeric_casts` carries no machine-applicable suggestion, so
    the 249 are stripped from the lint's own spans. Seven were wrapped in
    parentheses that existed only to hold the cast and had to go with it,
    or `unused_parens` (denied) rejects the result.

  * The lint is reported per macro *expansion*, and `check_ref_idx!` in
    both B-slice parsers is expanded with `$iref` as `i8` at one site and
    `i32` at another: dropping the cast there stops compiling. The
    `$listIdx` casts really were no-ops and go; the `$iref` one becomes
    `i32::from`, which is correct for both without a cast at all. An
    `#[allow]` on the macro would not have worked anyway — the level is
    resolved at the expansion site.

tests/common/mod.rs keeps two scoped allows on its convenience re-exports,
for the reason `dead_code` is allowed there: each consumer is its own
crate and pulls in one of them, or neither.

Clean on default / wide / scalar / wide+scalar, debug and release,
aarch64 and x86_64. Byte-identical: diffharness 847/847 over
st qp ps ltr sl dl mt def.

+++ rust/crates/openh264-rs/src/encoder/svc_set_mb_syn_cavlc.rs
@@ -12,7 +12,6 @@
 #![forbid(unsafe_code)]
 
 use crate::safe::bits::BsWriter;
-use crate::encoder::set_mb_syn_cabac::SCabacCtx;
 pub use crate::encoder::encoder_context::EWelsSliceType;
 pub use crate::encoder::encoder_context::SMVUnitXY;
 pub use crate::encoder::encoder_context::SDCTCoeff;
@@ -605,7 +604,7 @@ pub fn CheckBitstreamBuffer(
     let (pos, len) = (pBs.pos(), buf.len());
     debug_assert!(pos + 1 < len, "the writer is already at or past the buffer end");
 
-    if len < pos + 1 + MAX_MACROBLOCK_SIZE_IN_BYTE_x2 as usize {
+    if len < pos + 1 + MAX_MACROBLOCK_SIZE_IN_BYTE_x2 {
         return ENC_RETURN_VLCOVERFLOWFOUND;
     }
     ENC_RETURN_SUCCESS
